# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "教务系统",
        "description": "教务系统接口文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/teacher/add_bill": {
            "post": {
                "tags": [
                    "账单管理"
                ],
                "summary": "新增账单",
                "operationId": "37057c4818e7861805c382c841615c22",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "description": "课程id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teacher/list_bill": {
            "post": {
                "tags": [
                    "账单管理"
                ],
                "summary": "账单列表",
                "operationId": "7b800a349e2150125a129aa438885288",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "page_size",
                                    "page"
                                ],
                                "properties": {
                                    "page_size": {
                                        "description": "每页显示几条信息",
                                        "type": "integer"
                                    },
                                    "page": {
                                        "description": "当前显示第几页",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "course_id": {
                                                                "description": "课程id",
                                                                "type": "integer"
                                                            },
                                                            "created_at": {
                                                                "description": "创建时间",
                                                                "type": "string"
                                                            },
                                                            "status": {
                                                                "description": "账单状态",
                                                                "type": "integer"
                                                            },
                                                            "status_description": {
                                                                "description": "账单状态-中文描述",
                                                                "type": "string"
                                                            },
                                                            "course": {
                                                                "description": "课程详情",
                                                                "properties": {
                                                                    "id": {
                                                                        "description": "唯一标识",
                                                                        "type": "integer"
                                                                    },
                                                                    "date": {
                                                                        "description": "日期",
                                                                        "type": "string"
                                                                    },
                                                                    "course_name": {
                                                                        "description": "课程名称",
                                                                        "type": "string"
                                                                    },
                                                                    "cost": {
                                                                        "description": "费用",
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teacher/send_bill": {
            "get": {
                "tags": [
                    "账单管理"
                ],
                "summary": "发送账单",
                "operationId": "eec9ec2ebf0b51da360a00ce4f7a277f",
                "parameters": [
                    {
                        "name": "bill_id",
                        "in": "query",
                        "description": "账单id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization-Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}